remove warning

+++ cipher/prf.h
@@ -11,7 +11,7 @@ class PRF {
    public:
     PRF(){};
     ~PRF() {
-        for (int i = 0; i < pub_M.size(); i++) {
+        for (size_t i = 0; i < pub_M.size(); i++) {
             if (pub_M[i] != nullptr) {
                 delete[] pub_M[i];
             }
@@ -40,7 +40,7 @@ class PRF {
         Integer* tmp = new Integer[hmac.DIGLEN];
 
         A[0] = seed;
-        for (int i = 1; i < blks + 1; i++) {
+        for (size_t i = 1; i < blks + 1; i++) {
             hmac.hmac_sha256(tmp, A[i - 1]);
             hmac_calls_num++;
             concat(A[i], tmp, hmac.DIGLEN);
@@ -82,7 +82,7 @@ class PRF {
         uint32_t* tmpd = new uint32_t[hmac.DIGLEN];
 
         unsigned char* As = new unsigned char[32 + seedlen];
-        for (int i = 1; i < blks + 1; i++) {
+        for (size_t i = 1; i < blks + 1; i++) {
             hmac.opt_hmac_sha256(tmp, A[i - 1], hashlen[i - 1], reuse_in_hash_flag,
                                  reuse_out_hash_flag, zk_flag);
             hmac_calls_num++;
@@ -123,7 +123,7 @@ class PRF {
         res.bits.erase(res.bits.begin(),
                        res.bits.begin() + blks * (hmac.DIGLEN * hmac.WORDLEN) - bitlen);
 
-        for (int i = 0; i < blks + 1; i++) {
+        for (size_t i = 0; i < blks + 1; i++) {
             delete[] A[i];
         }
 
@@ -149,7 +149,7 @@ class PRF {
         reverse(rseed, rseed + seedlen);
         A[0] = Integer(8 * seedlen, rseed, ALICE);
 
-        for (int i = 1; i < blks + 1; i++) {
+        for (size_t i = 1; i < blks + 1; i++) {
             hmac.opt_rounds_hmac_sha256(tmp, A[i - 1], reuse_in_hash_flag,
                                         reuse_out_hash_flag);
             hmac_calls_num++;
@@ -228,7 +228,7 @@ class PRF {
     inline void prf_check(int party) {
         if (pub_M.size() != zk_sec_M.size())
             error("length of M is not consistent!\n");
-        for (int i = 0; i < pub_M.size(); i++)
+        for (size_t i = 0; i < pub_M.size(); i++)
             check_zero<IO>(zk_sec_M[i], pub_M[i], 8, party);
     }
 };
@@ -255,7 +255,7 @@ class PRFOffline {
         Integer* tmp = new Integer[hmac.DIGLEN];
         uint32_t* tmpd = new uint32_t[hmac.DIGLEN];
 
-        for (int i = 1; i < blks + 1; i++) {
+        for (size_t i = 1; i < blks + 1; i++) {
             hmac.opt_hmac_sha256(tmp, reuse_in_hash_flag, reuse_out_hash_flag);
             hmac_calls_num++;
 
@@ -289,7 +289,7 @@ class PRFOffline {
 
         Integer* tmp = new Integer[hmac.DIGLEN];
 
-        for (int i = 1; i < blks + 1; i++) {
+        for (size_t i = 1; i < blks + 1; i++) {
             hmac.opt_rounds_hmac_sha256(tmp, A[i - 1], reuse_in_hash_flag, reuse_out_hash_flag);
             hmac_calls_num++;
