+++ protocol/aead_izk.h
@@ -16,7 +16,7 @@ inline void itmac_hom_add_check(Integer& res, Integer& pre_res, int party, block
 
     if (party == BOB) {
         block delta = ((ZKVerifier<IO>*)(ProtocolExecution::prot_exec))->ostriple->delta;
-        for (int i = 0; i < pre_res.size(); i++) {
+        for (size_t i = 0; i < pre_res.size(); i++) {
             block tmp = set_bit(zero_block, i);
             block tmpx = tmp & blk;
             if (cmpBlock(&tmpx, &tmp, 1))
@@ -29,14 +29,14 @@ inline void itmac_hom_add_check(Integer& res, Integer& pre_res, int party, block
 // Implicitly homomorphic property and check zero of IT-MAC.
 template <typename IO>
 inline void itmac_hom_add_check(
-  Integer& res, Integer& pre_res, int party, const unsigned char* share, int len) {
+  Integer& res, Integer& pre_res, int party, const unsigned char* share, size_t len) {
     assert(pre_res.size() == (len * 8));
     assert(res.size() == (len * 8));
     if (party == BOB) {
         block delta = ((ZKVerifier<IO>*)(ProtocolExecution::prot_exec))->ostriple->delta;
         bool* data = new bool[len * 8];
         to_bool(data, share, len * 8);
-        for (int i = 0; i < len * 8; i++) {
+        for (size_t i = 0; i < len * 8; i++) {
             if (data[i])
                 pre_res[i].bit = pre_res[i].bit ^ delta;
         }
@@ -87,7 +87,7 @@ class AEAD_Proof {
 
     inline void gctr(Integer& res, size_t m) {
         Integer tmp(128, 0, PUBLIC);
-        for (int i = 0; i < m; i++) {
+        for (size_t i = 0; i < m; i++) {
             Integer content = nonce;
             tmp = computeAES_KS(expanded_key, content);
